Refactor Taskboard to just Board to avoid functional bias.
